Feature: Solana
Solana is a blockchain built for mass adoption. It's a high performance network that is utilized for a range of use cases, including finance, NFTs, payments, and gaming. Solana operates as a single global state machine, and is open, interoperable and decentralized. The goal of this branch is to implement functionality to generate wallets and sign transactions according to the Solana blockchain.

List of changes:
- implemented solana_address_encoder.dart for Base58 addresses used within Solana network
- created ed25519_derivator.dart containing key derivation functionality for base ED25519 curve
- created new coin configuration for Solana in bip44_coins_config.dart and slip44.dart
- added new HMAC key (ed25519 seed) to bip32_hmac_keys.dart as this value is used in wallet derivation