Refactor: Flutter dependencies removal
Previously, the cryptography_utils package had Flutter in its dependencies. Because of this, only Flutter projects could use this package, and consequently, Dart command line applications were not supported. Since Flutter is not necessary for using cryptography_utils, it was decided to remove it from the dependencies to enable testing without the need to use the Flutter framework.

List of changes:
- removed Flutter dependencies from pubspec.yaml
- removed ranged constraints from the packages version in pubspec.yaml to guarantee the use of their tested and verified version
- updated supported Dart version from ">=3.3.2" to ">=3.1.4" to allow projects using older Dart version to use cryptography_utils
- replaced "flutter_test" import into "test" in all tests
- removed Flutter-specific annotations like @optionalTypeArgs, @protected
- replaced compute() method usages with method from external package. Native dart language supports multithreading, but doesn't have predefined methods simplifying those operations.
- updated README.md to include information about Dart compatibility