Feature: Cosmos signing
This branch introduces the functionality for building and signing transactions in compliance with the Cosmos SDK. Currently, the only supported signing option is SIGN_MODE_DIRECT, as it's the only variant currently needed in the projects using "cryptography_utils" package. The other types, described in cosmos_sign_mode.dart, will be implemented as the project and its requirements evolve.

List of changes:
- created cosmos_signer.dart for creating Cosmos-compatible digital signatures using an ECDSA private key
- created cosmos_verifier.dart for verifying Cosmos-compatible digital signatures using an ECDSA public key
- created cosmos_signature.dart - a class storing digital signature used in Cosmos transactions
- created Dart representation of Cosmos protobuf objects ("transactions/cosmos") required to build Cosmos signing bytes and Cosmos transaction