Feature: Solana verifier
This branch includes code extracted from the initial implementation of Solana signer as a part of the solana-transactions branch. As Dominik mentioned in CR Point 8 of the aforementioned branch, EDDSA can never generate an invalid signature, so a solana signature verifier becomes fully optional. However, in case we ever need it for any reason, it may be quickly reimplemented with this branch.

List of changes:
- added solana_verifier.dart, providing the main functionality for verifying EDDSA signatures
- updated solana_signer.dart, making use of the SolanaVerifier